requirements.txt
absl-py==1.2.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
gast==0.3.3
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==2.10.0
idna==3.3
importlib-metadata==4.12.0
joblib==1.1.0
Keras-Preprocessing==1.1.2
littleutils==0.2.2
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
networkx==2.6.3
numpy==1.18.5
oauthlib==3.2.1
ogb==1.3.4
opt-einsum==3.3.0
outdated==0.2.1
pandas==1.3.5
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.2.1
rdkit==2022.3.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.4.1
six==1.16.0
spektral==1.2.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==2.0.1
threadpoolctl==3.1.0
torch==1.12.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
